refactor(cross-package): move getDeepActiveElement to @radix-ui/primitives
