--- packages/core/deep-active-element/README.md
@@ -1,3 +1,0 @@
-# `deep-active-element`
-
-This package contains a utility to get the currently focused element even across Shadow DOM boundaries.
