--- packages/core/deep-active-element/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "@radix-ui/deep-active-element",
-  "version": "1.0.0",
-  "license": "MIT",
-  "source": "./src/index.ts",
-  "main": "./src/index.ts",
-  "module": "./src/index.ts",
-  "publishConfig": {
-    "main": "./dist/index.js",
-    "module": "./dist/index.mjs",
-    "types": "./dist/index.d.ts",
-    "exports": {
-      ".": {
-        "import": {
-          "types": "./dist/index.d.mts",
-          "default": "./dist/index.mjs"
-        },
-        "require": {
-          "types": "./dist/index.d.ts",
-          "default": "./dist/index.js"
-        }
-      }
-    }
-  },
-  "files": [
-    "dist",
-    "README.md"
-  ],
-  "sideEffects": false,
-  "scripts": {
-    "lint": "oxlint --max-warnings 0 src",
-    "clean": "rm -rf dist",
-    "typecheck": "tsc --noEmit",
-    "build": "radix-build"
-  },
-  "devDependencies": {
-    "@repo/builder": "workspace:*",
-    "@repo/typescript-config": "workspace:*",
-    "typescript": "^5.7.3"
-  },
-  "homepage": "https://radix-ui.com/primitives",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/radix-ui/primitives.git"
-  },
-  "bugs": {
-    "url": "https://github.com/radix-ui/primitives/issues"
-  }
-}
