--- packages/core/deep-active-element/src/index.ts
@@ -1,1 +1,0 @@
-export { getDeepActiveElement } from './deep-active-element';
