--- packages/core/deep-active-element/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@repo/typescript-config/react-library.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "types": ["@repo/typescript-config/react-library"]
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist"]
-}
