+++ packages/react/password-toggle-field/package.json
@@ -36,7 +36,6 @@
     "build": "radix-build"
   },
   "dependencies": {
-    "@radix-ui/deep-active-element": "workspace:*",
     "@radix-ui/primitive": "workspace:*",
     "@radix-ui/react-compose-refs": "workspace:*",
     "@radix-ui/react-context": "workspace:*",
